refactor(screens): Settings & Week aufgeteilt (#71)
**Settings-Screen** (~265  ~71 Zeilen):
- ThemeSection, MaintenanceTools, ProfileSection, VersionInfo nach lib/features/settings/widgets/
- Screen nutzt nur noch importierte Widgets

**Week-Screen** (~432  ~147 Zeilen):
- WeekStats-Logik nach lib/features/week/logic/week_stats.dart
- WeekNavigationBar, WeekStatsCard, WeekExportCard, WeekAiAnalysisCard nach lib/features/week/widgets/
- Screen fokussiert auf Orchestrierung

Analyzer: sauber (flutter analyze)
Verhalten: unverändert (keine Business-Logik-Änderungen)
